Support Metrics Service Endpoint
This change introducse a new url endpoint for the metrics server.
This means the metric server can be deployed independently of the client server.

This commit also addresses a bug where the analytics service may not be initalized
if we failed to validate the API key.

The Target custom field has been renamed attributes