[FFM-1219]: Update clauses for processing segment match
What:
Changed how rules are processed to ensure they are using the correct
attribute.

Added exclude rule processing, to ensure we can exclude targets

Refacatored code into two methods for readabiltiy

Why:
The rules for determining if a target belonged to segment were looking
at the wrong attribute value.  This means the rules were broken.

Previously we didn't evaluate the exclude list, so it wasn't possible to exclude
a target from a group.