[FFM-1356]: Determine if a target should be included in a segment
What:
When a flag is being evaluated, that has been direcltly added to a segment,
we now call `isTargetIncludedBySegment()`  to ensure we check the
include, exclude and custom rules for matching targets to segments.

Why:
Previously we only checked if a target was matched to a segment with the include list.
We didn't evaluate the exclude list or custom rules.