[FFM-1358]: Analytics manager fails to start when debug is enabled
Currently the analytics manager enables a copy of the client API on startup,
however this is a static instance.
The problem when debug is enabled, that the manager didn't read the config, and so tries
to modify the static instance of the client API from a different thread (setting it to false, where as the main thread as set it to true).

This results in the manager throwing an exception and failing to start.
As a by product the isInitalize function will never return true.

What:
This change removes the second intialization of the static client API.  The client API is initalized from
the main constructor, so will already exist.
Instead of the analytics manager authenticating seperately it shares the same JWT as the the client API.