[FFM-1446]: Rules processing for integer and boolean types failed
What:
Use toString() instead of a cast to string

Why:
Target attributes are stored as there native type and converted to an Object
in java.
We were using a cast to convert them to a string representation to use in the
evaluator.
This failed silently as an exception was throw.  We need to use the toString()
method to correctly get the string representation of the attribute value.

Tests:
Added some unit tests to cover this