[FFM-2680]: Fix NPE when checking attributes map
 ### What
This change uses the Optional.ofNullable, which handles the null and
returns empty.

This change also removes the reflection when used to lookup the name
and identifier fields of a target.

 ### Why
When attempting to access an attribute in a map, which does not exist
a null was being passed to the Optional.of function.
This caused a NPE.

 ### Testing
A evaluation test was added to simulate this case.