FFM-5149 - Avoid using read timeout of 0 for SSE connections (#114)
What
SSE client should avoid using a read timeout of 0 since if the socket is not gracefully shutdown we have no way to detect that the socket is dead. Instead enforce a default read time out of 30 mins to force the connection to refresh and remove any stale socket.

Why
This will allow the client to self heal in the event of a broken connection and avoid a situtation where the client gets out of sync with the FF server.

Testing
Manual testing, checked client was able to recover its connection even after the socket was broken on purpose.