FFM-5251 - Merge FFM-5149 (flags out of sync) from 1.0.5.X to main (#117)
FFM-5251 - Merge FFM-5149 (flags out of sync) from 1.0.5.x to main

###What
SSE client should avoid using a read timeout of 0 since if the socket is not gracefully shutdown we have no way to detect that the socket is dead. Instead enforce a default read time out > 0 to force the connection to refresh and remove any stale socket. This change integrates the original bug fix FFM-5149 from 1.0.5.X into main. Main has previsouly been refactored and diverged so a 1:1 merge is not possible so a re-implementation has been done.

Additional changes:

- okhttp bumped from 3.9.1 to 4.9.0 to match 1.0.5.X version
- sseTimeout is now defined in HarnessConfig since Config was deprecated
- Fixed an issue where Poller would not restart on SSE disconnect due to the starttime not being set
- Added additional logging around socket events and ScheduledService states

###Why
This will allow the client to self heal in the event of a broken connection and avoid a situtation where the client gets out of sync with the FF server.

###Testing
Manual testing, checked client was able to recover its connection even after the socket was broken on purpose.