FFM-5306 - Update Java SDK to Match Variations for Prereqs With Identifier Instead of Values
What
When a feature depends on another prereq feature being true, the evaluation fails if the value and identifier are different values in the prereq.

Why
The checkPreRequisite() method incorrectly uses the value of the prereq instead of the identifier.

Testing
Manual + unit testing. Created flags with prereqs on app.harness.io, exported JSON into a local test case and added a new test called shouldCorrectlyEvaluatePrereqsIfIdAndValueDiffer()