[FFM-5471] - Target identifier set in the target builder function in the Java SDK is not used (#125)
What
Targets reported by Java SDK are incorrectly showing in the UI Target Management -> Targets as the global target

Why
We're sending the global target identifier in the targetData when we shouldn't be. Also global target is not correctly applied to metricsData when isGlobalTargetEnabled is set

Testing
Tested manually with sample app to confirm changes are reflected in the UI.
Also added a new unit test to capture and assert on the posted MetricsData object.