FFM-5928 - Fix java.lang.NoSuchMethodError caused by conflicting versions of kotlin-stdlib on the classpath
What
Exclude kotlin-stdlib being brought in by mapdb.
Minor fix to GettingStart example to stop it from closing the client too soon.

Why
okhttp 4.9.3 uses a newer version of the kotlin-stdlib which is conflicting with a version brought in by mapdb

Testing
Tested with GettingStarted java sample using same instructions as on the README