FFM-5111 - Java SDK - Polls non stop if /stream endpoint returns 5xx error
What
If a /stream endpoint returns a 501 HTTP error code (unimplemented) then the SSE library attempts reconnect endlessly

Why
The code is assuming 501 is a transient http error - so keeps retrying it rather than backing off for good

Testing
Added new mock web server dispatcher to mimic a 501 response when /stream is connected to.
New assertion were added to ensure SDK doesn't attempt to reconnect once when a 501 is returned.