FFM-7277 - Open the Evaluator and Query class/interfaces to allow usage (#146)
FFM-7277
This makes the Evaluator class and Query interface public, allowing users to instantiate their own instances.

Since the cache is already injectable, it can be reused with a users own instance of the Evaluator class. So if a user can instantiate their own instance of the Evaluator class, they can do evaluations on all flags, without triggering metric reporting.