[FFM-8688] - Handle accountID as an optional field in the JWT token
What
Adding new unit tests + harden the existing JWT parsing code to handle information which may be optional when an ff-proxy is in use.
If the info is not available the header will not be added. Also fixed env header to fall back to UUID when not present.

Why
The proxy doesn’t send accountID in the JWT token on auth, so we should make sure the code handles this correctly without bailing out when a proxy is in use.

Testing
New unit tests written + manual