[FFM-10816] - Remove metrics flush on map overflow
What
When too many evaluations are processed and with enough targets the internal metrics map we use will hit its limit very quickly. This causes the metrics to be flushed to the server.

Why
Flushing metrics is unbound and can result in more pressure on the server than necessary

Testing
Manual + unit