[FFM-10981] - Update Java SDK example to include a delay
What
The example on the main page is missing the sleep() statement that is included in the code example.
This updates the main page to match the code.

Why
Help avoid confusion around the life cycle of the SDK - the current example will exit immediately which is not what we want.