[FFM-11193] - Target v2: Adding support for AND/OR in clauses
What
Adding support for processing the new `GroupServingRule` in `Segment`. Also update API definition for OpenAPIGenerator plugin.
The default behaviour now is to check for the existence of `GroupServingRule` if found then we ignore the old `Rules` section.
Old `Rules` section will continue to work to aid transition if `GroupServingRule` is not yet populated.
Each entry in GroupServingRule is ORed while each list of clauses is ANDed. The GroupServingRule with the highest priority
used first to allow reordering on the backend.

Why
This give the customer more flexibility on how they configure their rules in the backend.

Testing
Initial unit tests for now + testgrid later